+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- **2024-04-17** 
+  - html:content:link:social-media:twitter - acknowledge www links
+  - html:content:link:social-media:linkedin - acknowledge de links
+  - social-media:twitter:<twitterHandle> - now  adding the Twitter handle
+
 - **2024-03-26**
   - hosting:tools:whmcs - added via WordPress theme
   - wordpress:plugin:avada - added via WordPress theme